Rename internal/* to pkg/*
This allows outside modules to import these packages.
