+++ pkg/bgp/testdata/open-3
