+++ pkg/controller/service.go
@@ -21,7 +21,7 @@ import (
 	"github.com/go-kit/kit/log"
 	v1 "k8s.io/api/core/v1"
 
-	"go.universe.tf/metallb/internal/allocator/k8salloc"
+	"go.universe.tf/metallb/pkg/allocator/k8salloc"
 )
 
 func (c *Controller) convergeBalancer(l log.Logger, key string, svc *v1.Service) bool {
