+++ pkg/k8s/stats.go
