+++ pkg/speaker/layer2_controller.go
@@ -21,8 +21,8 @@ import (
 	"sort"
 
 	"github.com/go-kit/kit/log"
-	"go.universe.tf/metallb/internal/config"
-	"go.universe.tf/metallb/internal/layer2"
+	"go.universe.tf/metallb/pkg/config"
+	"go.universe.tf/metallb/pkg/layer2"
 	v1 "k8s.io/api/core/v1"
 )
