+++ pkg/speaker/layer2_controller_test.go
@@ -6,8 +6,8 @@ import (
 	"sort"
 	"testing"
 
-	"go.universe.tf/metallb/internal/config"
-	"go.universe.tf/metallb/internal/k8s"
+	"go.universe.tf/metallb/pkg/config"
+	"go.universe.tf/metallb/pkg/k8s"
 
 	"github.com/go-kit/kit/log"
 	v1 "k8s.io/api/core/v1"
