+++ pkg/speaker/speaker.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"net"
 
-	"go.universe.tf/metallb/internal/bgp"
-	"go.universe.tf/metallb/internal/config"
-	"go.universe.tf/metallb/internal/k8s/types"
-	"go.universe.tf/metallb/internal/layer2"
+	"go.universe.tf/metallb/pkg/bgp"
+	"go.universe.tf/metallb/pkg/config"
+	"go.universe.tf/metallb/pkg/k8s/types"
+	"go.universe.tf/metallb/pkg/layer2"
 
 	gokitlog "github.com/go-kit/kit/log"
 	"github.com/prometheus/client_golang/prometheus"
