+++ pkg/speakerlist/speakerlist.go
@@ -19,7 +19,7 @@ import (
 	"sync"
 	"time"
 
-	"go.universe.tf/metallb/internal/k8s"
+	"go.universe.tf/metallb/pkg/k8s"
 
 	gokitlog "github.com/go-kit/kit/log"
 	"github.com/hashicorp/memberlist"
